rename: "create-vouchers" --> "vouchers"
